requirements_resume.txt
google-generativeai>=0.3.0
reportlab>=4.0